v2/requirements.txt
gpsd-py3
paho_mqtt>=2.0
